Move networkx into graph_libs
