+++ src/geff/graph_libs/_networkx.py
